Correction to serious page about bcrypt
With something as predictible in format as an IP address, recovering the
original IP address from the bcrypt hash is not impossible, only
infeasible. bcrypt hashes contain the salt, digest and number of rounds
used. Using this information you can attempt a brute force of the IP
address with the same salt and number of rounds. It will take a long
time, but it can be done.

The original text of the page is accurate, however it implies that
recovering an IP address from the bcrypt hash is impossible, which is
incorrect.