Create a thumbnail for each uploaded video
Also fixes opengraph support that was lost in the redesign.

We may want to change how this works. The converter for png images only
works with video files as the input. Might be worth tweaking it so that
we explicitly declare which input file type this converter is meant to
handle.

For what it's worth, we can convert other images into PNG files with
imagemagick like this: `convert input.ext output.png`