Add webm to converters and set as preferred source
Note that we still need to convert existing videos, which is probably a
day or two of hardcore processing on the server. This simply adds all
the support required to use webm for new videos.

We can push this after ffmpeg has been recompiled to support libvpx on
the server. Existing videos needn't have a webm source to work at that
point - the browser will fallback to one of the other available formats.