Move social links to top of drop box
This makes the /r/MediaCrush and @mediacru_sh links more apparent, but
maybe a little more intrusive. I think it looks fine.

This commit also makes minor responsiveness improvemens to the appearance of the drop
text on smaller layouts (namely by preventing the "Drag and drop or click to
upload" text from wrapping).

Closes #435